% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglik_rbm}
\alias{loglik_rbm}
\alias{loglik_rbm_approx}
\title{Compute Log-likelihood Value for RBM}
\usage{
loglik_rbm(w, b, c, dat)

loglik_rbm_approx(w, b, c, dat, nsamp = 100L, nstep = 10L, vr = TRUE)
}
\arguments{
\item{w}{Weight parameter of the RBM, of size \code{[m x n]}.}

\item{b}{Bias parameter for the visible units, of size \code{[m x 1]}.}

\item{c}{Bias parameter for the hidden units, of size \code{[n x 1]}.}

\item{dat}{The observed data, of size \code{[m x N]}.}

\item{nsamp}{Size of the Monte Carlo sample for approximation.}

\item{nstep}{Number of steps in the Gibbs sampler.}

\item{vr}{Whether to use variance reduction technique.}
}
\description{
The two functions compute the exact and approximate log-likelihood values for RBM, respectively.
}
\examples{
set.seed(123)
m = 10
n = 10
b = rnorm(m, sd = 0.1)
c = rnorm(n, sd = 0.1)
w = matrix(rnorm(m * n, sd = 1.0), m, n)

N = 100
dat = matrix(0, m, N)
v0 = rbinom(m, 1, 0.5)
for(i in 1:N)
{
    dat[, i] = rbm_sample_k(w, b, c, v0, k = 100)$v
}

loglik_rbm(w, b, c, dat)
loglik_rbm_approx(w, b, c, dat, nsamp = 100, nstep = 10)
loglik_rbm_approx(w, b, c, dat, nsamp = 100, nstep = 100)

}
