% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rbm_sample_k}
\alias{rbm_sample_k}
\alias{rbm_sample_tau}
\title{Sample from an RBM Model}
\usage{
rbm_sample_k(w, b, c, v0, k = 10L)

rbm_sample_tau(w, b, c, v0, min_steps = 10L, max_steps = 100L,
  verbose = FALSE)
}
\arguments{
\item{w}{Weight parameter of the RBM, of size \code{[m x n]}.}

\item{b}{Bias parameter for the visible units, of size \code{[m x 1]}.}

\item{c}{Bias parameter for the hidden units, of size \code{[n x 1]}.}

\item{v0}{The initial value for Gibbs sampling, of size \code{[m x 1]}.}

\item{k}{Number of steps in the Gibbs sampler.}

\item{min_steps}{Minimum number of steps in the Gibbs sampler.}

\item{max_steps}{Maximum number of steps in the Gibbs sampler.}

\item{verbose}{Whether to print algorithmic information.}
}
\description{
\code{rbm_sample_k()} uses Gibbs sampling with a fixed-size chain.
\code{rbm_sample_tau()} uses the unbiased Gibbs sampling with a chain path.
}
\examples{
set.seed(123)
m = 10
n = 10
b = rnorm(m, sd = 0.1)
c = rnorm(n, sd = 0.1)
w = matrix(rnorm(m * n, sd = 1.0), m, n)
v0 = rbinom(m, 1, 0.5)

rbm_sample_k(w, b, c, v0, k = 100)
rbm_sample_tau(w, b, c, v0, min_steps = 1, max_steps = 100, verbose = TRUE)

}
